// types/quizAttempt/enums.ts
export enum QuizAttemptMessage {
  SUCCESS_SUBMIT = "Nộp bài quiz thành công",
  FAIL_SUBMIT = "Nộp bài quiz thất bại",
  
  SUCCESS_GET = "Lấy thông tin kết quả quiz thành công",
  FAIL_GET = "Lấy thông tin kết quả quiz thất bại",
  
  ATTEMPT_NOT_FOUND = "Không tìm thấy kết quả làm bài",
  QUIZ_NOT_FOUND = "Không tìm thấy quiz",
  
  INVALID_ANSWERS = "Câu trả lời không hợp lệ",
  MISSING_ANSWERS = "Chưa trả lời đủ câu hỏi",
  
  CONGRATULATIONS_PASSED = "Chúc mừng! Bạn đã vượt qua bài quiz",
  TRY_AGAIN = "Bạn chưa đạt điểm. Hãy thử lại!"
}